+++ DataLayer/ItemsDBO.h
@@ -17,11 +17,11 @@ INT32 DeserializeItem(Item* i, MemoryStream * data);
 
 Item *CreateItem(Item * existing, UID playerId, sql::Connection * sql);
 Item *CreateItemFromTemplate(UINT32 iTemplateIndex, UID playerId, sql::Connection * sql);
-Item *CreateItem(UINT64 itemId, UID playerId, sql::Connection * sql);
+Item *CreateItem(UINT32 itemId, UID playerId, sql::Connection * sql);
 
 Item* CreateNonDbItem(Item * existing);
 Item* CreateNonDbItemFromTemplate(UINT32 iTemplateIndex, UINT32 stackCount);
-Item* CreateNonDbItem(UINT64 itemId, UINT32 stack);
+Item* CreateNonDbItem(UINT32 itemId, UINT32 stack);
 
 INT32 UpdateItem(Item * i, sql::Connection * conn);
 INT32 DeleteItem(UID id, sql::Connection * conn);
